Instancio 1 objeto de cada clase en JuegoHP.java

+++ src/app/artefactos/CapaInvisibilidad.java
@@ -6,7 +6,7 @@ public class CapaInvisibilidad extends Artefacto implements IReliquiaMuerte {
 
     // CONSTRUCTOR
     
-    public CapaInvisibilidad(String nombre, int amplificadorDeDanio, int amplificadorDeCuracion) {
+    public CapaInvisibilidad(String nombre, double amplificadorDeDanio, double amplificadorDeCuracion) {
         this.setNombre(nombre);
         this.setAmplificadorDeDanio(amplificadorDeDanio);
         this.setAmplificadorDeCuracion(amplificadorDeCuracion);
